Common hal_gpio implementation for all stm32 mcus
