+++ hw/mcu/stm/stm32_common/src/hal_gpio.c
@@ -17,20 +17,9 @@
  * under the License.
  */
 
-/*
- * Currently no support for
- * - interrupts
- * - pin output speed
- */
-
-#include "hal_gpio_stm32.h"
 #include "hal/hal_gpio.h"
 #include "mcu/cmsis_nvic.h"
-#include "stm32f3xx.h"
-#include "stm32f3xx_hal_cortex.h"
-#include "stm32f3xx_hal_gpio.h"
-#include "stm32f3xx_hal_rcc.h"
-#include "mcu/stm32f3xx_mynewt_hal.h"
+#include "mcu/stm32_hal.h"
 #include <assert.h>
 
  /* XXX: Notes
@@ -57,7 +46,7 @@
 /*
  * GPIO pin mapping
  *
- * The stm32F3xx processors have 16 gpio pins per port. We map the logical pin
+ * The stm32Fxxx processors have 16 gpio pins per port. We map the logical pin
  * numbers (from 0 to N) in the following manner:
  *      Port A: PA0-PA15 map to pins 0 - 15.
  *      Port B: PB0-PB15 map to pins 16 - 31.
@@ -71,9 +60,82 @@
  *      Ex: PD11 = (3 * 16) + 11 = 59.
  *          PA0 = (0 * 16) + 0 = 0
  */
-#define HAL_GPIO_PIN_NUM(pin)   ((pin) & 0x0F)
-#define HAL_GPIO_PORT(pin)      (((pin) >> 4) & 0x0F)
-#define HAL_GPIO_PIN(pin)       (1 << HAL_GPIO_PIN_NUM(pin))
+#define GPIO_INDEX(pin)     ((pin) & 0x0F)
+#define GPIO_PORT(pin)      (((pin) >> 4) & 0x0F)
+#define GPIO_MASK(pin)      (1 << GPIO_INDEX(pin))
+
+#ifndef GPIOA
+#   define GPIOA    0
+#endif
+#ifndef GPIOB
+#   define GPIOB    0
+#endif
+#ifndef GPIOC
+#   define GPIOC    0
+#endif
+#ifndef GPIOD
+#   define GPIOD    0
+#endif
+#ifndef GPIOE
+#   define GPIOE    0
+#endif
+#ifndef GPIOF
+#   define GPIOF    0
+#endif
+#ifndef GPIOG
+#   define GPIOG    0
+#endif
+#ifndef GPIOH
+#   define GPIOH    0
+#endif
+#ifndef GPIOI
+#   define GPIOI    0
+#endif
+#ifndef GPIOJ
+#   define GPIOJ    0
+#endif
+#ifndef GPIOK
+#   define GPIOK    0
+#endif
+
+
+#if defined GPIOK_BASE
+#   define  HAL_GPIO_PORT_COUNT   (11)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE,GPIOF,GPIOG,GPIOH,GPIOI,GPIOJ,GPIOK
+#elif defined GPIOJ_BASE
+#   define  HAL_GPIO_PORT_COUNT   (10)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE,GPIOF,GPIOG,GPIOH,GPIOI,GPIOJ
+#elif defined GPIOI_BASE
+#   define  HAL_GPIO_PORT_COUNT   (9)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE,GPIOF,GPIOG,GPIOH,GPIOI
+#elif defined GPIOH_BASE
+#   define  HAL_GPIO_PORT_COUNT   (8)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE,GPIOF,GPIOG,GPIOH
+#elif defined GPIOG_BASE
+#   define  HAL_GPIO_PORT_COUNT   (7)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE,GPIOF,GPIOG
+#elif defined GPIOF_BASE
+#   define  HAL_GPIO_PORT_COUNT   (6)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE,GPIOF
+#elif defined GPIOE_BASE
+#   define  HAL_GPIO_PORT_COUNT   (5)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD,GPIOE
+#elif defined GPIOD_BASE
+#   define  HAL_GPIO_PORT_COUNT   (4)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC,GPIOD
+#elif defined GPIOC_BASE
+#   define  HAL_GPIO_PORT_COUNT   (3)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB,GPIOC
+#elif defined GPIOB_BASE
+#   define  HAL_GPIO_PORT_COUNT   (2)
+#   define  HAL_GPIO_PORT_LIST    GPIOA,GPIOB
+#elif defined GPIOA_BASE
+#   define  HAL_GPIO_PORT_COUNT   (1)
+#   define  HAL_GPIO_PORT_LIST    GPIOA
+#else
+#   error "No GPIO ports found - MCU not supported!"
+#endif
+
 
 /* Port index to port address map */
 static GPIO_TypeDef * const portmap[HAL_GPIO_PORT_COUNT] =
@@ -81,87 +143,122 @@ static GPIO_TypeDef * const portmap[HAL_GPIO_PORT_COUNT] =
     HAL_GPIO_PORT_LIST
 };
 
-/* Storage for GPIO ISRS. */
-struct hal_gpio_irq_isr_t
+/* Storage for GPIO callbacks. */
+struct gpio_irq_obj
 {
-    hal_gpio_irq_handler_t  isr;
-    void                    *arg;
-    volatile uint32_t       invoked;
-    volatile uint32_t       dropped;
+    void *arg;
+    hal_gpio_irq_handler_t isr;
 };
-struct hal_gpio_irq_isr_t hal_gpio_irq[16];
 
-void hal_gpio_ext_irq_handler(uint16_t index)
+static struct gpio_irq_obj gpio_irq_handlers[16];
+
+struct ext_irqs
+{
+    volatile uint32_t irq0;
+    volatile uint32_t irq1;
+    volatile uint32_t irq2;
+    volatile uint32_t irq3;
+    volatile uint32_t irq4;
+    volatile uint32_t irq9_5;
+    volatile uint32_t irq15_10;
+};
+struct ext_irqs ext_irq_counts;
+
+/**
+ * ext irq handler
+ *
+ * Handles the gpio interrupt attached to a gpio pin.
+ *
+ * @param index
+ */
+static void
+ext_irq_handler(int index)
 {
     uint32_t mask;
 
-    mask = HAL_GPIO_PIN(index);
+    mask = 1 << index;
     if (__HAL_GPIO_EXTI_GET_IT(mask) != RESET) {
         __HAL_GPIO_EXTI_CLEAR_IT(mask);
-      if (hal_gpio_irq[index].isr) {
-          ++hal_gpio_irq[index].invoked;
-          hal_gpio_irq[index].isr(hal_gpio_irq[index].arg);
-      } else {
-          ++hal_gpio_irq[index].dropped;
-      }
+        gpio_irq_handlers[index].isr(gpio_irq_handlers[index].arg);
     }
 }
 
-
 /* External interrupt 0 */
 static void
-hal_gpio_ext_irq0(void)
+ext_irq0(void)
 {
-    hal_gpio_ext_irq_handler(0);
+    ++ext_irq_counts.irq0;
+    ext_irq_handler(0);
 }
 
 /* External interrupt 1 */
 static void
-hal_gpio_ext_irq1(void)
+ext_irq1(void)
 {
-    hal_gpio_ext_irq_handler(1);
+    ++ext_irq_counts.irq1;
+    ext_irq_handler(1);
 }
 
 /* External interrupt 2 */
 static void
-hal_gpio_ext_irq2(void)
+ext_irq2(void)
 {
-    hal_gpio_ext_irq_handler(2);
+    ++ext_irq_counts.irq2;
+    ext_irq_handler(2);
 }
 
 /* External interrupt 3 */
 static void
-hal_gpio_ext_irq3(void)
+ext_irq3(void)
 {
-    hal_gpio_ext_irq_handler(3);
+    ++ext_irq_counts.irq3;
+    ext_irq_handler(3);
 }
 
-/* External interrupt 4 */
+/**
+ * ext irq4
+ *
+ *  External interrupt handler for external interrupt 4.
+ *
+ */
 static void
-hal_gpio_ext_irq4(void)
+ext_irq4(void)
 {
-    hal_gpio_ext_irq_handler(4);
+    ++ext_irq_counts.irq4;
+    ext_irq_handler(4);
 }
 
-/* External interrupt 9 through 5 */
+/**
+ * ext irq9_5
+ *
+ *  External interrupt handler for irqs 9 through 5.
+ *
+ */
 static void
-hal_gpio_ext_irq9_5(void)
+ext_irq9_5(void)
 {
     int index;
 
+    ++ext_irq_counts.irq9_5;
     for (index = 5; index <= 9; ++index) {
-        hal_gpio_ext_irq_handler(index);
+        ext_irq_handler(index);
     }
 }
 
-/* External interrupt 15 through 10 */
+/**
+ * ext irq15_10
+ *
+ *  External interrupt handler for irqs 15 through 10.
+ *
+ */
 static void
-hal_gpio_ext_irq15_10(void)
+ext_irq15_10(void)
 {
     int index;
 
+    ++ext_irq_counts.irq15_10;
     for (index = 10; index <= 15; ++index) {
-        hal_gpio_ext_irq_handler(index);
+        ext_irq_handler(index);
     }
 }
 
@@ -176,79 +273,59 @@ static void
 hal_gpio_clk_enable(uint32_t port_idx)
 {
     switch (port_idx) {
+#if defined GPIOA_BASE
         case 0:
-            if (!__HAL_RCC_GPIOA_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOA_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOA_CLK_ENABLE();
             break;
-#if HAL_GPIO_PORT_COUNT > 1
+#endif
+#if defined GPIOB_BASE
         case 1:
-            if (!__HAL_RCC_GPIOB_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOB_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOB_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 2
+#if defined GPIOC_BASE
         case 2:
-            if (!__HAL_RCC_GPIOC_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOC_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOC_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 3
+#if defined GPIOD_BASE
         case 3:
-            if (!__HAL_RCC_GPIOD_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOD_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOD_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 4 && defined GPIOE_BASE
+#if defined GPIOE_BASE
         case 4:
-            if (!__HAL_RCC_GPIOE_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOE_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOE_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 5
+#if defined GPIOF_BASE
         case 5:
-            if (!__HAL_RCC_GPIOF_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOF_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOF_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 6
+#if defined GPIOG_BASE
         case 6:
-            if (!__HAL_RCC_GPIOG_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOG_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOG_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 7
+#if defined GPIOH_BASE
         case 7:
-            if (!__HAL_RCC_GPIOH_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOH_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOH_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 8
+#if defined GPIOI_BASE
         case 8:
-            if (!__HAL_RCC_GPIOI_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOI_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOI_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 9
+#if defined GPIOJ_BASE
         case 9:
-            if (!__HAL_RCC_GPIOJ_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOJ_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOJ_CLK_ENABLE();
             break;
 #endif
-#if HAL_GPIO_PORT_COUNT > 10
+#if defined GPIOK_BASE
         case 10:
-            if (!__HAL_RCC_GPIOK_IS_CLK_ENABLED()) {
-                __HAL_RCC_GPIOK_CLK_ENABLE();
-            }
+            __HAL_RCC_GPIOK_CLK_ENABLE();
             break;
 #endif
         default:
@@ -273,7 +350,7 @@ hal_gpio_pin_to_irq(int pin)
     int index;
     IRQn_Type irqn;
 
-    index = HAL_GPIO_PIN_NUM(pin);
+    index = GPIO_INDEX(pin);
     if (index <= 4) {
         irqn = EXTI0_IRQn + index;
     } else if (index <=  9) {
@@ -285,32 +362,36 @@ hal_gpio_pin_to_irq(int pin)
     return irqn;
 }
 
+#if MYNEWT_VAL(MCU_STM32F3)
+#   define EXTI2_IRQn   EXTI2_TSC_IRQn
+#endif
+
 static void
 hal_gpio_set_nvic(IRQn_Type irqn)
 {
-    void (*isr)(void);
+    uint32_t isr;
 
     switch (irqn) {
     case EXTI0_IRQn:
-        isr = hal_gpio_ext_irq0;
+        isr = (uint32_t)&ext_irq0;
         break;
     case EXTI1_IRQn:
-        isr = hal_gpio_ext_irq1;
+        isr = (uint32_t)&ext_irq1;
         break;
-    case EXTI2_TSC_IRQn:
-        isr = hal_gpio_ext_irq2;
+    case EXTI2_IRQn:
+        isr = (uint32_t)&ext_irq2;
         break;
     case EXTI3_IRQn:
-        isr = hal_gpio_ext_irq3;
+        isr = (uint32_t)&ext_irq3;
         break;
     case EXTI4_IRQn:
-        isr = hal_gpio_ext_irq4;
+        isr = (uint32_t)&ext_irq4;
         break;
     case EXTI9_5_IRQn:
-        isr = hal_gpio_ext_irq9_5;
+        isr = (uint32_t)&ext_irq9_5;
         break;
     case EXTI15_10_IRQn:
-        isr = hal_gpio_ext_irq15_10;
+        isr = (uint32_t)&ext_irq15_10;
         break;
     default:
         assert(0);
@@ -318,40 +399,10 @@ hal_gpio_set_nvic(IRQn_Type irqn)
     }
 
     /* Set isr in vector table if not yet set */
-    if (NVIC_GetVector(irqn) != (uint32_t)isr) {
-        NVIC_SetVector(irqn, (uint32_t)isr);
+    if (NVIC_GetVector(irqn) != isr) {
+        NVIC_SetVector(irqn, isr);
         NVIC_EnableIRQ(irqn);
     }
-}
-
-
-static int
-hal_gpio_init_stm32_int(int pin, GPIO_InitTypeDef *cfg, GPIO_PinState *state)
-{
-    int port;
-    uint32_t mcu_gpio_pin;
-
-    /* Is this a valid pin? */
-    port = HAL_GPIO_PORT(pin);
-    if (port >= HAL_GPIO_PORT_COUNT) {
-        return -1;
-    }
-
-    mcu_gpio_pin = HAL_GPIO_PIN(pin);
-    cfg->Pin = mcu_gpio_pin;
-
-    /* Enable the GPIO clock */
-    hal_gpio_clk_enable(port);
-
-    /* Write initial state if required */
-    if (NULL != state) {
-        HAL_GPIO_WritePin(portmap[port], mcu_gpio_pin, *state);
-    }
-
-    /* Initialize pin setting proper mode */
-    HAL_GPIO_Init(portmap[port], cfg);
-
-    return 0;
 }
 
 /**
@@ -367,8 +418,27 @@ hal_gpio_init_stm32_int(int pin, GPIO_InitTypeDef *cfg, GPIO_PinState *state)
 int
 hal_gpio_init_stm(int pin, GPIO_InitTypeDef *cfg)
 {
-  return hal_gpio_init_stm32_int(pin, cfg, NULL);
+    int port;
+    uint32_t mcu_pin_mask;
+
+    /* Is this a valid pin? */
+    port = GPIO_PORT(pin);
+    if (port >= HAL_GPIO_PORT_COUNT) {
+        return -1;
+    }
+
+    mcu_pin_mask = GPIO_MASK(pin);
+    cfg->Pin = mcu_pin_mask;
+
+    /* Enable the GPIO clock */
+    hal_gpio_clk_enable(port);
+
+    /* Initialize pin as an input, setting proper mode */
+    HAL_GPIO_Init(portmap[port], cfg);
+
+    return 0;
 }
+
 /**
  * hal gpio deinit
  *
@@ -383,16 +453,16 @@ int
 hal_gpio_deinit_stm(int pin, GPIO_InitTypeDef *cfg)
 {
     int port;
-    uint32_t mcu_gpio_pin;
+    uint32_t mcu_pin_mask;
 
     /* Is this a valid pin? */
-    port = HAL_GPIO_PORT(pin);
+    port = GPIO_PORT(pin);
     if (port >= HAL_GPIO_PORT_COUNT) {
         return -1;
     }
 
-    mcu_gpio_pin = HAL_GPIO_PIN(pin);
-    cfg->Pin = mcu_gpio_pin;
+    mcu_pin_mask = GPIO_MASK(pin);
+    cfg->Pin = mcu_pin_mask;
 
     /* Initialize pin as an input, setting proper mode */
     HAL_GPIO_DeInit(portmap[port], cfg->Pin);
@@ -437,36 +507,36 @@ hal_gpio_init_in(int pin, hal_gpio_pull_t pull)
 int hal_gpio_init_out(int pin, int val)
 {
     GPIO_InitTypeDef cfg;
-    GPIO_PinState state;
+    int port;
 
-    cfg.Mode  = GPIO_MODE_OUTPUT_PP;
-    cfg.Pull  = GPIO_NOPULL;
-    cfg.Speed = GPIO_SPEED_FREQ_HIGH;
-
-    state = val ? GPIO_PIN_SET : GPIO_PIN_RESET;
-
-    return hal_gpio_init_stm32_int(pin, &cfg, &state);
-}
-
-/**
- * gpio enumeration translation
- *
- * Convert hal pull types to mcu pull type
- *
- * @param pull
- *
- * @return uint32_t
- */
-static uint32_t
-hal_gpio_pull_to_stm_pull(enum hal_gpio_pull pull)
-{
-    switch (pull) {
-        case HAL_GPIO_PULL_NONE:  return GPIO_NOPULL;
-        case HAL_GPIO_PULL_UP:    return GPIO_PULLUP;
-        case HAL_GPIO_PULL_DOWN:  return GPIO_PULLDOWN;
+    /* Is this a valid pin? */
+    port = GPIO_PORT(pin);
+    if (port >= HAL_GPIO_PORT_COUNT) {
+        return -1;
     }
-    /* never reached */
-    return GPIO_NOPULL;
+
+    /* Enable the GPIO clock */
+    hal_gpio_clk_enable(port);
+
+    /* Write initial output value */
+    hal_gpio_write(pin, val);
+
+    cfg.Pin = GPIO_MASK(pin);
+    cfg.Mode = GPIO_MODE_OUTPUT_PP;
+    cfg.Pull = GPIO_NOPULL;
+#if (defined GPIO_SPEED_FREQ_VERY_HIGH)
+    cfg.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
+#elif (defined GPIO_SPEED_HIGH)
+    cfg.Speed = GPIO_SPEED_HIGH;
+#else
+    cfg.Speed = GPIO_SPEED_FREQ_HIGH;
+#endif
+    cfg.Alternate = 0;
+
+    /* Initialize pin as an output, setting proper mode */
+    HAL_GPIO_Init(portmap[port], &cfg);
+
+    return 0;
 }
 
 /**
@@ -477,14 +547,24 @@ hal_gpio_pull_to_stm_pull(enum hal_gpio_pull pull)
 int
 hal_gpio_init_af(int pin, uint8_t af_type, enum hal_gpio_pull pull, uint8_t od)
 {
-    GPIO_InitTypeDef cfg;
+    GPIO_InitTypeDef gpio;
 
-    cfg.Mode      = od ? GPIO_MODE_AF_OD : GPIO_MODE_AF_PP;
-    cfg.Speed     = GPIO_SPEED_FREQ_HIGH;
-    cfg.Pull      = hal_gpio_pull_to_stm_pull(pull);
-    cfg.Alternate = af_type;
+    if (!od) {
+        gpio.Mode = GPIO_MODE_AF_PP;
+    } else {
+        gpio.Mode = GPIO_MODE_AF_OD;
+    }
+#if (defined GPIO_SPEED_FREQ_VERY_HIGH)
+    gpio.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
+#elif (defined GPIO_SPEED_HIGH)
+    gpio.Speed = GPIO_SPEED_HIGH;
+#else
+    gpio.Speed = GPIO_SPEED_FREQ_HIGH;
+#endif
+    gpio.Pull = pull;
+    gpio.Alternate = af_type;
 
-    return hal_gpio_init_stm(pin, &cfg);
+    return hal_gpio_init_stm(pin, &gpio);
 }
 
 /**
@@ -498,14 +578,19 @@ hal_gpio_init_af(int pin, uint8_t af_type, enum hal_gpio_pull pull, uint8_t od)
 void hal_gpio_write(int pin, int val)
 {
     int port;
-    uint32_t mcu_gpio_pin;
+    uint32_t mcu_pin_mask;
     GPIO_PinState state;
 
-    port         = HAL_GPIO_PORT(pin);
-    mcu_gpio_pin = HAL_GPIO_PIN(pin);
-    state        = val ? GPIO_PIN_SET : GPIO_PIN_RESET;
+    port = GPIO_PORT(pin);
+    mcu_pin_mask = GPIO_MASK(pin);
 
-    HAL_GPIO_WritePin(portmap[port], mcu_gpio_pin, state);
+    if (val) {
+        state = GPIO_PIN_SET;
+    } else {
+        state = GPIO_PIN_RESET;
+    }
+
+    HAL_GPIO_WritePin(portmap[port], mcu_pin_mask, state);
 }
 
 /**
@@ -520,11 +605,11 @@ void hal_gpio_write(int pin, int val)
 int hal_gpio_read(int pin)
 {
     int port;
-    uint32_t mcu_gpio_pin;
+    uint32_t mcu_pin_mask;
 
-    port = HAL_GPIO_PORT(pin);
-    mcu_gpio_pin = HAL_GPIO_PIN(pin);
-    return HAL_GPIO_ReadPin(portmap[port], mcu_gpio_pin);
+    port = GPIO_PORT(pin);
+    mcu_pin_mask = GPIO_MASK(pin);
+    return HAL_GPIO_ReadPin(portmap[port], mcu_pin_mask);
 }
 
 /**
@@ -565,7 +650,7 @@ hal_gpio_irq_init(int pin, hal_gpio_irq_handler_t handler, void *arg,
     int index;
     uint32_t pin_mask;
     uint32_t mode;
-    GPIO_InitTypeDef cfg;
+    GPIO_InitTypeDef init_cfg;
 
     /* Configure the gpio for an external interrupt */
     rc = 0;
@@ -597,19 +682,18 @@ hal_gpio_irq_init(int pin, hal_gpio_irq_handler_t handler, void *arg,
     if (!rc) {
         /* Disable interrupt and clear any pending */
         hal_gpio_irq_disable(pin);
-        pin_mask = HAL_GPIO_PIN(pin);
+        pin_mask = GPIO_MASK(pin);
         __HAL_GPIO_EXTI_CLEAR_FLAG(pin_mask);
 
         /* Set the gpio irq handler */
-        index = HAL_GPIO_PIN_NUM(pin);
-        hal_gpio_irq[index].isr = handler;
-        hal_gpio_irq[index].arg = arg;
-        hal_gpio_irq[index].invoked = 0;
+        index = GPIO_INDEX(pin);
+        gpio_irq_handlers[index].isr = handler;
+        gpio_irq_handlers[index].arg = arg;
 
         /* Configure the GPIO */
-        cfg.Mode = mode;
-        cfg.Pull = pull;
-        rc = hal_gpio_init_stm(pin, &cfg);
+        init_cfg.Mode = mode;
+        init_cfg.Pull = pull;
+        rc = hal_gpio_init_stm(pin, &init_cfg);
         if (!rc) {
             /* Enable interrupt vector in NVIC */
             irqn = hal_gpio_pin_to_irq(pin);
@@ -640,13 +724,13 @@ hal_gpio_irq_release(int pin)
     hal_gpio_irq_disable(pin);
 
     /* Clear any pending interrupts */
-    pin_mask = HAL_GPIO_PIN(pin);
+    pin_mask = GPIO_MASK(pin);
     __HAL_GPIO_EXTI_CLEAR_FLAG(pin_mask);
 
     /* Clear out the irq handler */
-    index = HAL_GPIO_PIN_NUM(pin);
-    hal_gpio_irq[index].arg = NULL;
-    hal_gpio_irq[index].isr = NULL;
+    index = GPIO_INDEX(pin);
+    gpio_irq_handlers[index].arg = NULL;
+    gpio_irq_handlers[index].isr = NULL;
 }
 
 /**
@@ -662,7 +746,7 @@ hal_gpio_irq_enable(int pin)
     uint32_t ctx;
     uint32_t mask;
 
-    mask = HAL_GPIO_PIN(pin);
+    mask = GPIO_MASK(pin);
 
     __HAL_DISABLE_INTERRUPTS(ctx);
     EXTI->IMR |= mask;
@@ -681,7 +765,7 @@ hal_gpio_irq_disable(int pin)
     uint32_t ctx;
     uint32_t mask;
 
-    mask = HAL_GPIO_PIN(pin);
+    mask = GPIO_MASK(pin);
     __HAL_DISABLE_INTERRUPTS(ctx);
     EXTI->IMR &= ~mask;
     __HAL_ENABLE_INTERRUPTS(ctx);
